이재원

+++ Final/src/kh/c/five/model/RegiDto.java
@@ -12,7 +12,7 @@ public class RegiDto implements Serializable {
 	private String rs_category; //한식 중식 일식 etc
 	private String rs_menu; //식당메뉴
 	private String rs_time; //영업시간
-	private int rs_rating; //별점
+	private double rs_rating; //별점
 	private int rs_readcount; //조회수
 	private String rs_parking;
 	//private int rs_parking; //주차장시설이 있다 ->1 주차장시설이 없다 ->0
@@ -23,7 +23,7 @@ public class RegiDto implements Serializable {
 	public RegiDto() {}
 
 	public RegiDto(int seq, String rs_name, String rs_address1, String rs_address2, String rs_picture,
-			String rs_category, String rs_menu, String rs_time, int rs_rating, int rs_readcount, String rs_parking,
+			String rs_category, String rs_menu, String rs_time, double rs_rating, int rs_readcount, String rs_parking,
 			String rs_price, String rs_phone, String rs_keyword) {
 		super();
 		this.seq = seq;
@@ -106,11 +106,11 @@ public class RegiDto implements Serializable {
 		this.rs_time = rs_time;
 	}
 
-	public int getRs_rating() {
+	public double getRs_rating() {
 		return rs_rating;
 	}
 
-	public void setRs_rating(int rs_rating) {
+	public void setRs_rating(double rs_rating) {
 		this.rs_rating = rs_rating;
 	}
